Rust: Refactor Modules
